refactor: move storage to dedicated pkg
